Move fixtures to conftest.py
